third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-once-expected.txt

CONSOLE ERROR: line 23: Uncaught SyntaxError: Unexpected token }
This is a testharness.js-based test.
PASS Invalid uncompiled raw handlers should only be compiled once. 
Harness: the test ran to completion.
